rename members components
