--- libs/organization/src/lib/components/member-invitation/member-invitation.component.html
@@ -1,14 +1,0 @@
-<mat-card>
-  <h3>Pending members ({{ invitations.length }})</h3>
-  <mat-list *ngIf="invitations.length; else empty" role="list" class="bf-list">
-    <span matListItem *ngFor="let invitation of invitations" fxLayout="row" fxLayoutAlign="space-between center">
-      <p>{{ invitation.user.email }}</p>
-      <button *ngIf="isAdmin" (click)="declined.emit(invitation)" mat-icon-button>
-        <mat-icon svgIcon="refuse"></mat-icon>
-      </button>
-    </span>
-  </mat-list>
-  <ng-template #empty>
-    <span>No pending invitations</span>
-  </ng-template>
-</mat-card>
