--- libs/organization/src/lib/components/member-invitation/member-invitation.component.scss
@@ -1,3 +1,0 @@
-p {
-  padding-left: 10px;
-}
