+++ libs/organization/src/lib/components/member-pending/member-pending.component.scss
@@ -1,6 +1,3 @@
-.bold {
-  font-weight: bold;
-  text-transform: capitalize;
-  margin-right: 10px;
-  margin-left: 10px;
+p {
+  padding-left: 10px;
 }
