+++ libs/organization/src/lib/components/member-request/member-request.component.html
@@ -1,0 +1,26 @@
+<mat-card>
+  <h3>New Member Requests</h3>
+  <bf-table-filter
+    [source]="users"
+    [columns]="invitationColumns"
+    [initialColumns]="initialColumns"
+    showPaginator="true" >
+    <ng-template colRef="name" let-name>
+      <b>{{ name }}</b>
+    </ng-template>
+    <ng-template colRef="surname" let-surname>
+      <b>{{ surname }}</b>
+    </ng-template>
+    <ng-template colRef="email" let-email>
+      {{ email }}
+    </ng-template>
+    <ng-template colRef="uid" let-uid>
+      <button mat-icon-button (click)="accepted.emit(findInvitation(uid))">
+        <mat-icon svgIcon="accept_member"></mat-icon>
+      </button>
+      <button mat-icon-button (click)="declined.emit(findInvitation(uid))">
+        <mat-icon svgIcon="refuse"></mat-icon>
+      </button>
+    </ng-template>
+  </bf-table-filter>
+</mat-card>
