+++ libs/organization/src/lib/components/member-request/member-request.component.scss
@@ -1,0 +1,6 @@
+.bold {
+  font-weight: bold;
+  text-transform: capitalize;
+  margin-right: 10px;
+  margin-left: 10px;
+}
