+++ libs/organization/src/lib/organization.module.ts
@@ -38,10 +38,10 @@ import { MatTooltipModule } from '@angular/material';
 // Components
 import { OrganizationDisplayComponent } from './components/organization-display/organization-display.component';
 import { MemberAddComponent } from './components/member-add/member-add.component';
-import { MemberPendingComponent } from './components/member-pending/member-pending.component';
-import { MemberInvitationComponent } from './components/member-invitation/member-invitation.component';
 import { MemberFormRoleComponent } from './components/member-form-role/member-form-role.component';
 import { TableFilterModule } from '@blockframes/ui/list/table-filter/table-filter.module';
+import { MemberRequestComponent } from './components/member-request/member-request.component';
+import { MemberPendingComponent } from './components/member-pending/member-pending.component';
 
 // Pages
 import { MemberRepertoryComponent } from './components/member-repertory/member-repertory.component';
@@ -122,9 +122,9 @@ export const organizationRoutes: Routes = [
     RouterModule.forChild(organizationRoutes)
   ],
   declarations: [
-    MemberPendingComponent,
+    MemberRequestComponent,
     MemberFormRoleComponent,
-    MemberInvitationComponent,
+    MemberPendingComponent,
     OrganizationDisplayComponent,
     MemberEditableComponent,
     MemberRepertoryComponent,
